[Std/util] Improve DEFALIST.
Now STD::DEFALIST always generates a (disabled) rewrite rule that concludes
ALISTP from the recognizer of the alist type being defined.

Supporting the disabling of the theorem has necessitated extending
DEF-GENERIC-RULE (in [books]/std/lists/abstract.lisp) with a new :DISABLE
option, which is therefore implicitly available to DEF-ALISTP-RULE.

A few books had to be adapted by removing from them the theorem that is now
generated (the theorems there were apparently not redundant with the generated
ones).

Thanks to Sol Swords for explaining how to support the generation of disabled
theorems in STD::DEFALIST (i.e. the approach described above).