[ATC] Restrict use of function calls.
This is in preparation for supporting C code with side effects. We plan to allow
non-pure expressions only in certain places (e.g. initializers of declarations,
or expression statements), so that we can keep most of the expressions pure.

C function calls, while not necessarily non-pure, will be regarded as non-pure,
which simplifies the formulation of the semantics.

The user doc has been updated to describe the revised representation of C code
in ACL2.  In particular, there is now a new notion of allowed pure non-boolean
term, which excludes function calls of ACL2 functions that represent C
functions, and of allowed non-boolean term, which does.

A test for function calls has been adapted to adhere to the new restrictions.