[PFCS] Add a new library for PFCS.
PFCS (Prime Field Constraint Systems) are a notion, introduced and formalized in
this library, that generalizes R1CS (Rank-1 Constraint Systems) and likely other
systems of constraints for zero-knowledge proofs.

This PFCS library complements the R1CS library, adding support for more modular
and compositional proofs and for richer transformations.