[ATC] Generalize C dynamic semantics.
Now the execution of function calls also returns a possibly updated computation
state -- currently it is never updated, but this generalization is necessary to
support general side effects.

The returned computation state is obtained by popping the frame from the
computation state resulting from the execution of the function's body.