Tweak a call of COMMENT in prover output pertaining to unknown warrants.
Thanks to J for pointing out and discussing this issue.

The last example in :doc comment showed this:

    (IMPLIES
     (NOT (APPLY$-WARRANT-BAR))
     (EQUAL (HIDE (COMMENT "Call failed because the warrant for BAR is false"
                           (EV$ '(BAR Y) '((Y . 3)))))
            3))

That was correct, but in general the warrant in such a case is merely
unknown, not necessarily false.  That can happen when forcing is
disabled (otherwise the warrant would be assumed true by forcing it).

So now that checkpoint looks as follows:

    (IMPLIES
     (NOT (APPLY$-WARRANT-BAR))
     (EQUAL (HIDE (COMMENT "Call failed because the warrant for BAR is not known to be true"
                           (EV$ '(BAR Y) '((Y . 3)))))
            3))

Additional explanation has been added to :doc comment.