[ATC] Add a lemma to support symbolic execution.
This seems needed for some examples that use assignment, but it may not have to
do specifically with assignments, but rather with a certain code patterns used
in functions involved in defining the dynamic semantics of assignments. In any
case, this suggests that the rules for symbolic execution should be rationalized
more.