[ATC] Add a new tutorial page.
This page explains the generation of C code with multiple functions, the
restrictions on their relative ordering and (non-)recursion, and the current
generation of a single file (i.e. translation unit, which is the same as a
preprocessing translation unit as far as the C code generated by ATC is
concerned).