[ATC] Improve proofs generated by ATC.
Now the functional correctness proofs are carried out by symbolic execution, via
a carefully selected set of rules that is designed to be necessary adn
sufficient to carry out the proofs efficiently and robustly.

Indeed, the new approach results in much faster proofs, and works for all the
current tests except one. This may be due to some heuristic on lambda expansion
getting in the way; the issue is being investigated. This test has been
temporarily commented out.