[ATC] Generate additional theorems.
For every function, now we generate a more general correctness theorem, which
talks about the execution of the function in any computation state, not only in
the initial computation state.

The new theorems are only local for now, and independent from the correctness
theorems that were generated before (and that are still being
generated). However, we plan to use the new more general theorems to prove the
less general ones.

Perhaps more importantly, the more general theorems are a step towards more
modular and compositional proofs, in which theorems about callees are used to
prove theorems about callers.

Some rules have to be added to the symbolic execution theory in order to
accommodate thes slightly more complex proofs.