[ATC] Modularize proof generation function-wise.
Now the correctness proof of each function makes use of the correctness theorems
of the called functions, as opposed to having the symbolic execution unfold all
functions calls. This makes the generated proofs much more scalable and
efficient.

It was also necessary to generate, for each function, a theorem asserting that
it returns a C value; the already generated theorems asserting that each
function does not return an error did not suffice for these new function-wise
proofs to work.