[ATC] Extend abstract syntax with declarators.
Introduce an explicit notion of declarators, which are part of declarations,
according to the C standard. For now we only support restricted forms of
declarations and declarators, but this extension paves the way to supporting
more forms.

Adapt pretty-printer, static semantics, dynamic semantics, proof support
(including limit theorems), and code generation to the extension.