[ATC] Extend declarators.
Besides declarators of the form <identifier>, also allow declarators of the form
*<identifier>, i.e. for pointers. These are also commonly used for arrays in
function parameters, which is our motivation for adding them now, namely to
support the generation of C functions that take arrays as arguments.

Adapt the pretty-printer to print the new declarators.

Adapt static and dynamic semantics to reject these new declarators for
now. Support will be added later.