[ATC] Update formulation of static correctness theorem.
It turns out that it was still assuming that CHECK-TRANSUNIT returned a boolean,
when instead it was changes some time ago to return either :WELLFORMED or an
error (the latter with richer information). So now the formulation has been
updated accordingly.

All the instances of this theorem for the tests still go through. This is an
easy theorem, proved by straight execution since it calls an executable function
on a named constant. It simply asserts that the generated C code compiles.