[ATC] Generalize RUN-FUN.
This is the top-level of the C dynamic semantics, which initializes a
computation state and runs a specified C function on specified arguments.  It
now takes a heap as argument, which is put into the initial computation
state. This is necessary to properly generate proofs for code that uses arrays.

The plan is to eventually remove this function, and instead export theorems
about EXEC-FUN. But for now, this is a smaller change that will support the
generation of proofs for arrays,