[ATC] Extend the C static semantics.
Add support for all the integer constants. This involves, based on the unsigned
and type suffixes, and the base (10, 8, 16), finding, in the applicable list,
the smallest type where the value fits; the constant is illegal if it does not
fit in any type. (We restrict things to the standar integer types; we have no
extended integer types.)