Slightly generalise two lemmas
This generalisation, applied to both these lemmas, is somewhat limited
because ACL2 is often able to draw the obvious conclusion when two
number-valued terms satisfy <= and >= at the same time. Sometimes ACL2
isn't, though, and that makes this generalisation worthwhile. It
brings in a case-split arising from the definitions of <= and >=
respectively, but that has not proven to be a problem anywhere in the
regression suite.