[ATC] Add some integer operations on lower-rank types.
These are versions of other operations that we already have for signed and
unsigned ints, longs, and long longs. The new versions are for signed and
unsigned chars and shorts. But these are only a few of the integer operations:
most of them do not operate directly on these lower-rank types, due to the
integer promotions and to the usual arithmetic conversions.