[ATC] Generate more general shift op targets.
These are ACL2 functions that are targets of refinement for generating C
code. We generate, for each C integer type of rank int or higher, a function for
left shifts and a function for right shifts: while the first argument (which
represents the first operand) has the associated C type, the second argument is
an ACL2 integer, for which the function targets <TYPE>-INTEGER-VALUE can be used
to turn any (same or different) C integer type into the shift distance.

We maintain the older shift targets temporarily, and define them in terms of the
more general ones, but we will delete the older ones once we move everything
else over to using the newwer shift targets.

Adapt a guard proof in one of the tests to the modified definition of the older
shift targets.