[ATC] Rename some functions of the C model.
SINT01 => SINT-FROM-BOOLEAN
<TYPE>-NONZEROP => BOOLEAN-FROM-<TYPE>

The new names more clearly expresses that the first function turns an ACL2
boolean into a C int value (0 or 1) and that the other functions turn C values
(of various types) to ACL2 booleans.

This matches the naming structures of other functions of the C model, in
particular the <type1>-from-<type2> integer conversions.

Thanks to Eric Smith for a discussion that led to this improvement.