[Zcash] Prove that Pedersen encoding <.> is not 0.
This requires a lemma about the loop function saying that it yields values
outside an interval that depends on the J argument. Then the main theorem
follows from the fact that the <.> function calls the loop function with J = 1
and thus being outside the (open!) interval from -1 to 1 is equivalent to not
being 0.