[ATC] Speed up the symbolic execution proofs.
Exclude rules that handle error situations that are not expected to occur in the
generated proofs. This idea came about by looking at the accumulated persistence
results on one of the slow-running proofs described in recent commits. Excluding
this rule cuts the time of one of this proofs from about 48 seconds to about 36
seconds -- just a measurement for each.