Second version of child stobjs of abstract stobj.  Also fixed bug when there are two updates in a stobj-let.
There are many issues remaining, more testing, documentation, etc.
NOTE: There was one "everything" regression failure, which of course
needs to be addressed.

This improves the first version in several ways, including relaxation
of a couple of checks (input signature match with :logic, allow
aliases for stobj-let for the case that concrete array accessors agree
provided the indices are distinct).

See books/system/tests/nested-stobj-two-updates.lisp for an example of
the bug when there are two updates in a stobj-let.

books/centaur/aignet/from-hons-aig.cert