Fixed soundness bug in functions that print to strings.  Use evisc-tuple argument in fmt ~f and ~F directives.
Quoting :doc note-8-4 (where the first of these is modified from a
corresponding existing paragraph by adding "the evisc-tuple
argument"):

  For [fmt] directives ~f and ~F, ACL2 now uses the alist component of
  the evisc-tuple argument and the global [evisc-table].  Previously
  it used these only for directives ~x, ~y, ~X, and ~Y (and
  deprecated directives ~p etc.; see [fmt]).  Thanks to Eric Smith
  for a query leading to this enhancement.

  A soundness bug was fixed in the functions that print to strings; see
  [printing-to-strings].  The bug was a dependence of the result on
  the ACL2 [state], even though state is not an argument to these
  functions.  More specifically, the dependence was on the
  [current-package] and the global [evisc-table].

I changed two books to deal with the bug fix and have contacted the
authors.  A "regression-everything" passed.

Thanks to Eric Smith for encouraging me to look for more than one bug
in printing-to-strings; I initially found the current-package problem
but then looked further and found the evisc-table problem.