Third version of child stobjs of abstract stobj.
This passed an "everything" regression.  But there are some remaining
issues, as well as more tests and documentation to write.

This improves the second version in several ways:

- Avoid proof failures due to non-exec wrapper in abstract stobj
  primitives (which has been eliminated).

- Avoid going into an illegal state when interrupting a stobj-let form
  for an abstract stobj, if that form does no updating.

- There is less aggressive marking of invariant-risk on abstract stobj
  primitives; now that is based simply on the :exec function for the
  primitive.

- (Kind of unrelated) Improved the HIDE term's error explanation when
  logic evaluation is incomplete due to a call of non-exec.

Note that when a stobj-let does only a single update, there might not
be a need to enter an :illegal-state when the computation is
interrupted.  I might look into this at some point.