Fourth version of child stobjs of abstract stobj.
This passed an "everything" regression.  But as before, there are some
remaining issues, as well as more tests and documentation to write.

This improves the third version by weakening the check for aliasing in
stobj-let so that read-only access duplication is allowed for child
stobjs.  Except, this hasn't been handled yet for duplicate array
indices.

Also, improved some relevant tests, in part by adding to them.