[ATC] Add missing check.
There was a missing check that statement terms transforming one or more
variables must in fact end with those variables on all paths. ATC was correctly
recognizing and handling the situations in which that was the case, but was
neglecting to enforce that, prematurely falling back to the treating the term as
a C-valued one, which is appropriate only if the statement term is transforming
no variables. The generated C code was thus nonsensical.

Thanks to Eric Smith for attempting to generate C code from an ACL2 function
that is not rejected by the added check in ATC.