[ATC] Fix bug.
This is the handling of LETs in certain cases. In representing C assignments
inside a term that transforms variables other than the one being assigned, the
variable must be in the same C scope as the assignment. Otherwise, C code that
following the transforming code could reference a modified value, while the
corresponding ACL2 code would still reference the old value.