[ATC] Fix a bug.
This is related to the one I fixed yesterday. While the validation check I added
yesterday prevented the unsound modification of a variable from an outer scope,
today's fix restricts the kinds of transforming terms in LET and MV-LET bindings
to be IFs, so that their branches push new C scopes and the validation check I
added yesterday applies. Without this restriction, the transforming term could
be just a LET (sequence) that does not push new C scopes, therefore passing the
validation check added yesterday. The restriction added today does not really
restrict the representable C code in ACL2: transforming terms should always be
IFs, because if they are not they do not need to transforming terms and can
instead be sequenced with the preceding LETs.

Thanks to Eric Smith for reporting this second bug.