[ATC] Add a warning to user doc.
Explain how, in some (rare) cases, the well-formedness theorem may fail due to
ATC's current reliance on ACL2 guard verification to ensure the type correctness
of the C code. Since the former is stronger than the latter, ATC may generate
type-incorrect C code from ACL2 code that is unreachable under the guards and
that violates the guard conditions (which will pass guard verification because
the unreachability under the guards effectively supplies a NIL assumption).

Thanks to Jim McDonald for creating a test that exposed the issue, and thanks to
Jim McDonald and Eric Smith for useful discussion.