[ATC] Start generating events for programs with loops.
It is no longer necessary to have :PROOFS NIL when some target functions are
recursive, i.e. when the C program includes loops. ATC automatically avoids
generating theorems for programs with loops; the fact that ATC does not print
the theorems as part of its result conveys the absence of those theorems. This
is a temporary situation, in which we can gradually increase the events
generated for loops. In particular, now the DEFCONST for the translation unit
AST is always generated, i.e. also for programs including loops.