[bigmem] new library that defines a big array (2^64 bytes)
This library implements the idea in the following paper using nested
and abstract stobjs, which leads to a simpler implementation.

Warren A. Hunt, Jr. and Matt Kaufmann. A Formal Model of a Large
Memory that Supports Efficient Execution. In Proceedings of the 12th
International Conference on Formal Methods in Computer-Aided Design
(FMCAD 2012, Cambrige, UK, October 22-25), 2012

Example application: bigmem can be used as a child stobj to define a
field representing the memory (<= 2^64 bytes) in a parent stobj that
models some machine's state.

TODO: Documentation.