[ATC] Restructure some aspects of the symbolic execution.
This commit introduces a new representation for the computation states during
symbolic execution. It does not change the definition of computation states: it
merely introduces functions that update the computation state in a way that the
updates are easily visible and traversable when these new functions are kept
disabled during symbolic execution. Theorems are added that say how the existing
functions that manipulate the computation state transform this new
representation of the computation state.

Among other things, this paves the way to handling symbolic execution proofs for
code with side effects that apply to parts of the computation state that whose
structure is not completely known during the symbolic execution, as needed to
prove sufficiently general forms of the theorems that can be used in
incremental, modular proofs.