Quit when acl2-customization file has an error.  Allow :guard in :assumptions argument of SIMPLIFY transformation and improved some related error-checking.  Tweaked :doc certify-book (macro that returns error-triple).  Tweaked :pl error message when supplied macro name, to suggest term.  Changed "name" to "<name>" in output of :help.  Fixed a low-level bug in the SIMPLIFY transformation (handling assumptions and hyp-fn).
Quoting :doc note-8-4:

  When an error occurs while loading an [ACL2-customization] file, ACL2
  quits with exit code 1.  Thanks to Eric Smith for suggesting the
  quit and to Eric McCarthy for suggesting exit code 1 in that case.

Quoting :doc note-8-4-books:

  The apt::simplify transformation has been improved by allowing the
  use of keyword :guard for the :assumptions, not only as the entire
  value of :assumptions as before, but also as a member of a
  specified list of assumptions.

Regarding that improvement, a bit more error-checking is done for the
value of :assumptions.

Tweaked :doc certify-book to say that it is a macro that returns an
error-triple.  Thanks to Eric McCarthy for this suggestion.

Tweaked :pl so that if it is supplied a symbol that names a macro that is not a
macro-alias for a function symbol, then an extra sentence is printed suggesting
to use :pl on a term.  Thanks to Eric Smith for this suggestion.

Quoting a comment in (defxdoc note-8-4 ...):
; Tweaked the output of :help to be about "<name>" rather than "name".  Thanks
; to Alessandro Coglio for suggesting this change.

Fixed a low-level bug in the SIMPLIFY transformation's handling of
assumptions and hyp-fn (function fix-assumptions**).