Removed books/system/doc/stobj-fields-of-abstract-stobjs.txt.
I've dealt with all the pending items in that file, either by acting
on them or by deciding that it's best not to do so.  Well, with one
exception: I still plan to try to avoid macroexpansion for forms that
should be handled by the system function, translate11 -- notably
stobj-let forms with guard checks on possible duplicate array indices.
I'll probably do that soon.