[ATC] Remove functions and theorems no longer needed.
After some aspects of the generated proofs were changed recently, these are no
longer needed. The theorems themselves would not hurt to keep, but the functions
that these theorems are based on are variants of the mutually recursive
execution functions that must be manually modified every time the execution
functions are modified, so overall it is more burden than benefit.