Improved "make proofs" to make it more complete.  Improved :doc for get-event, logical-defun, and IO?.
Quoting :doc note-8-4:

  ACL2 now does a more thorough job of doing proofs when ``make
  proofs'' is executed.  In particular, proofs were formerly skipped
  but are now performed for [defun] forms containing the explicit
  [xargs] [declaration], :mode :logic, and for [defthm] events
  evaluated with [default-defun-mode] :logic outside the so-called
  ``pass 2 files''.

In addition to successfully running a fresh "regression-everything"
(as usual), successfully ran "make proofs" both for a regular build
and for a #+acl2-devel build, and successfully ran "make devel-check".

Improved :doc system-utilities to clarify that get-event gives the
most recent event with the given name and logical-defun may be
preferred for function symbols.  Thanks to Alessandro Coglio for
feedback leading to these changes.

Improved :doc system-utilities to clarify that it's evaluation, not
just printing, that can be suppressed with IO?.  Thanks to Eric Smith
for an email exchange leading to this fix.