[ATC] Extend input validation.
Ensure that the recursive target functions are singly (not mutually) recursive.

This is actually indirectly already checked when checking that the recursive
functions have the form required to represent C loops. However, here we add a
preliminary check at input processing time (vs. code generation time) to alert
the user right away of something that may be amiss.