Fixed cert_param comment(s).
Thanks to Sol for analyzing syntactic errors in an earlier attempt
(not committed) that had led me to use two cert_param comments.  Now
there's just one, which is nicer.  The "uses-acl2r" is probably not
necessary because of included books, but if it's to be there then it
should be "uses_acl2r", so I made that fix too.

The other changes are just trailing whitespace deletion (due to how
my emacs saves files).