[ATC] Calculate and store limits for loop functions.
Now all the target ACL2 functions, recursive and non-recursive, have limit terms
associated to them in the function table. If there are no recursive functions,
these terms are all quoted constants. In the presence of recursion, the
recursive function, and all its direct and indirect callers, will have a limit
term that in general depends on each function's parameters.