[ATC] Add experimental support for array writes.
When the :EXPERIMENTAL input includes :ARRAY-WRITES, now ATC recognizes LET
bindings of the form

(LET ((<ARRAY> (C::UCHAR-ARRAY-WRITE-SINT <ARRAY> <INDEX> <ELEMENT>))) ...)

and translates them to C array assignments. It also recognizes, in non-loop
functions, ending terms of the form

(MV <RETVAL> <ARRAY> ...)

where the first argument represents the return value (which must be a pure
C-valued expression in this case) and where the other arguments are arrays, and
translates this to a C return statement for <RETVAL>, ignoring the <ARRAY>
arguments because they are implicitly modified by side effects in C.

However, there are currently no checks for stobj-like single-threadedness, so
the generated code may be incorrect if those checks are not satisfied.

Proof generation is automatically disabled when :EXPERIMENTAL includes
:ARRAY-WRITES, except for the static well-formedness theorem.