[ATC] Weaken a check.
As part of the experimental support for array writes, a check that is performed
by ATC and is never expected to fail in the normal case (i.e. without array
write support) is now failing in some cases. Thus, the check has been relaxed to
only apply if :ARRAY-WRITES is not part of the :EXPERIMENTAL option of ATC.

Thanks to Karthik Nukala for reporting the occurrence of the error.