[ATC] Extend symbolic execution theory.
Add rules that bridge the gap between (1) integer promotions and usual
arithmetic conversions possibly leading to two sequential conversions in the
deep embedding and (2) single conversions in the shallow embedding for
operations on different types.

Thanks to Karthik Nukala for trying ATC on an example that exhibited a proof
failure due to the lack of these rules.