[ATC] Extend the model of C with the void type.
This involves extending abstract syntax, pretty-printer, static semantics, and
dynamic semantics, and also adapting some ATC code to deal with the possibility
of this extra type.

ATC does not yet generate C code that uses void, which will be supported in
upcoming commits. This commit just extends the model of C.