[ATC] Extend the statement generation code.
Accompany the terms being processed with alists from variables to terms that
capture the MV-LET and LET bindings that lead to the terms. The "real" term is
the one obtained by applying the alist substitution. This is needed to properly
instantiate the subterms of the limits that correspond to function calls: we
need to know the "real" arguments of the function calls, not just the terms that
occur in the calls.