Sped up certification of books/nonstd/circles/area-of-a-circle/*.lisp.  Added licenses for those two books and parent directory's book.
Certification of the two (ACL2(r)-only) books was reduced from over an
hour to under two minutes.  This was accomplished mostly by using
accumulated-persistence, but also by interrupting after
(set-debugger-enable t) and looking at rules mentioned on the stack.

;;; Before the changes:
area-of-a-circle-1.cert (378.920s)
area-of-a-circle-2.cert (3497.887s)
[Total:]
3870.013u 6.982s 1:04:37.98 99.9%	0+0k 283216+1336io 205pf+0w

;;; After the changes:
area-of-a-circle-1.cert (77.511s)
area-of-a-circle-2.cert (26.836s)
[Total:]
104.272u 0.501s 1:45.62 99.1%	0+0k 200+1176io 0pf+0w

....................

Here is the email thread with Ruben Gamboa authorization the addition
of licenses.

Matt Kaufmann <matthew.j.kaufmann@gmail.com>    Sat, Jul 31, 2021 at 2:08 PM
To: Ruben Gamboa <ruben@uwyo.edu>, Jagadish Bapanapally <jagadishb285@gmail.com>
Hi Ruben and Jagadish,

I just now noticed that none of the books has a license
(only copyright):

books/nonstd/circles/circumference-of-a-circle.lisp
books/nonstd/circles/area-of-a-circle/area-of-a-circle-1.lisp
books/nonstd/circles/area-of-a-circle/area-of-a-circle-2.lisp

A simple way to fix that would be for you to authorize me
to add the following line immediately under the copyright.
(That's what I typically do; others do differently.)

; License: A 3-clause BSD license.  See the LICENSE file distributed with ACL2.

Do you want me to do that? ....

..........

Ruben A Gamboa <ruben@uwyo.edu>    Sat, Jul 31, 2021 at 2:28 PM
To: Matt Kaufmann <matthew.j.kaufmann@gmail.com>
Yes absolutely!  Thanks.

..........