[ATC] Improve model of C arrays.
Use the newly introduced <TYPE>-ARRAY-LENGTH functions to define the predicates
that say whether an index is within an array's bounds. This enables the
replacement, in guards and similar situations, of terms of the form

(LEN (<TYPE>-ARRAY->ELEMENTS ...))

with the shorter and more abstract

(<TYPE>-ARRAY-LENGTH ...)

where the latter should probably be kept always disabled in guard proofs.

Adapt some array tests to the change.