[Yul] Improve the abstract syntax.
Model blocks as wrappers of statement lists, as opposed to just statement
lists. Even though the two are isomorphic, the wrapper lets us treat blocks more
independently, which seems useful given that in Yul objects code is represented
as blocks.

This also lets us more clearly process a block, in the static semantics, by
first collecting all the function definitions in it. This will be in fact added
in an upcoming commit.