[Yul] Extend the static semantics.
Now the checking of blocks returns not only an updated variable table, but also
an updated function table.

The functions defined in a block are now collected prior to checking the
statements of the block.

The initializtion block of the 'for' statement is now treated specially as
prescribed by the Yul documentation, namely by applying its updated variable and
function tables to the other components of the statement.