[Yul] Extend static semantics.
Add a flag that says whether we are in a loop's body block. (Despite its
apparent analogy to the flag saying whether we are in a loop's initialization
block, the two flags are actually handled differently, as explained in the
documentation.)

Use the flag to check that 'break' and 'continue' statements only occur in loop
bodies.