[Yul] Generalize the abstract syntax a bit.
Make the default block of switch statements optional, as in fact required by the
grammar.

While from the point of view of execution we could just use an empty block
(which does nothing) to represent an absent default block, we prefer to be
closer to the concrete syntax, as done for other parts of the abstract syntax.