[Yul] Extend static semantics.
Finish the coverage of 'switch' statements.

At this point a first draft of the static semantics is complete.

With this extension, the DEFINES at the end of static-semantics.lisp failed at
the generated MAKE-FLAG, which is unexpected. Adding :NORMALIZE NIL to the
CHECK-STATEMENT funtion avoids the problem: thanks to Matt Kaufmann for
suggesting this workaround after examining the problem.