[ATC] Remove some integer size rules from the symbolic execution theory.
Yesterday's addition of all the rules about integer sizes should have been
restricted to only the ones relating types of different (un)signedness.  The
ones relating types with the same (un)signedness may cause equality
substitutions that prevent the application of some of the other rules that are
meant to apply.

Thanks to Jim McDonald for an example of a proof failure that was caused by this
issue.