Changed APT::SIMPLIFY to print prover output when verify-guards fails and updated :DOC simplify-failure accordingly.  Also made a few minor tweaks to the SIMPLIFY implementation.  Fixed a minor bug in the fancy-use proof-builder command.
Eric Smith requested this verify-guards proof output so that he can
look at checkpoints to come up with helpful rules, without having to
run SIMPLIFY again.  Output is suppressed however when SIMPLIFY is
called with :print nil.

The tweaks were mostly to comments, but included removal of a
with-output call that I believe to be unnecessary.

The bug fix for acl2-pc::fancy-use avoids generating (orelse cmd nil).