[Zcash] Explicate a theorem.
This was used locally to prove the non-zeroness of the Pedersen encoding
function, but it turns out to be useful also for the upcoming proof of
injectivity of the Pedersen encoding function (which is a bit different from the
one in the Zcash specification, and rather it is more suited to the ACL2
formalization). Since that theorem involves a local auxiliary predicate, we
create a theorem that does not mention that predicate, but rather its expansion.