[DEFINES] Suppress return-value theorems in :program mode.
Fixes issue #1309.  An existing code comment mentioned this issue but seems wrongheaded, as exemplified by the issue filed by Eric McCarthy.  It seems perfectly reasonable to supply return value types in :program mode, either simply as documentation or because some code is in :program mode only temporarily.  In that case, this commit ensures that defines will not fail.

Normally, generated theorems would be simply skipped in :program mode.  The failure here is caused by the use of defthm-flag to prove them, since the corresponding make-flag is skipped in :program mode.

Standard disclaimers apply: My head is not really in this code.