Delete questionable files ending in *.bak and *.orig
Currently, after `make basic` is run and the books directory is ready
for installation, five files ending in *.bak or *.orig remain which
should not be included in the final installation. All five files reside
in the directory books/workshops/1999/ivy/ivy-v2/ivy-sources/ and
should instead be cleaned. This has already been done in Debian [1]
and Gentoo [2].

[1] https://debian.pkgs.org/9/debian-main-amd64/acl2-emacs_7.2dfsg-3_all.deb.html
[2] https://github.com/gentoo/sci/blob/a9a6dc87759ff69f93105d2b85b5cd90755380e5/sci-mathematics/acl2/acl2-8.4.ebuild#L31