[Zcash] Prove two theorems.
One just says that a Jubjub point satisfies the curve equation. It is expressed
in terms of the Jubjub projection functions, so that we can avoid expanding
everything into the functions from the twisted Edwards curve library.

The other theorem says that a Jubjub point cannot have abscissa 1. This is a key
property needed to prove a theorem from the Zcash specification saying that a
particular value is not in the image of Pedersen hash.