[ATC] Improve and fix proof generation.
This commit combines two orthogonal changes.

One change consists in enabling NOT in certain proofs. The motivation for this
is that we found a case in which, due to a clausification heuristics, a proof
fails even though logically speaking it should succeed by simple propositional
reasoning. Experiments on a scaled-down example showed that enabling NOT got
past that issue. Enabling NOT in the proofs generated by ATC caused no
regressions and made the failing example work.

The other change simplifies the form of the loop theorem, at the same time
fixing a bug in proof generation due to an incomplete variable substitution.
This latter failure occurred in the same example mentioned above, after the
enabling of NOT got past the initial issue.

Thanks to Eric Smith for running ATC on an example that resulted in the issue(s)
above, and also for working out the scaled-down example. Thanks to Eric Smith
and Matt Kaufmann for useful discussion on this topic.