Modified stobj-table access in stobj-let to pass in creater call instead of applying fixer.  Also made a few minor changes, including error messages.
Thanks to Sol Swords for pointing out that the initial version of
stobj-tables, which required using a fixer for accesses in stobj-let
bindings, could invalidate read-over-write theorems.  Thanks to both
Rob Sumners and Sol for helpful discussions towards this revision.

The :doc remains to be updated.

;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
;;; Here is a longer discussion of these changes: ;;;
;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;

This change to stobj-tables eliminates the need for hypotheses in
read-over-write theorems -- a problem identified by Sol Swords -- in
the direction agreed to with Rob Sumners and Sol.  See section
"Read-over-write issue" in
books/system/tests/stobj-table-tests-input.lsp.

The big change at the user level is that stobj-let bindings no longer
use a stobj-fixer.

The documentation still needs to be updated, but for now you can
probably get what you need by looking at
books/system/tests/stobj-table-tests-input.lsp.

I considered eliminating stobj-fixers altogether, but I did actually
find a new use for them: they are wrapped around stobj-table accessor
calls before generating guard proof obligations.  There is a comment
about this in books/system/tests/stobj-table-tests-input.lsp, just
above (verify-guards update-1).  There is also a use for stobj-fixers
in section "Reasoning about stobj recognizers" at the end of that
file.  The basic idea there is that now, since it is no longer
logically valid (even though it's an execution invariant) that a
stobj-table access produces a stobj -- we had to give that up by
eliminating the stobj-fixer application to that result, to get the
simpler read-over-write theorems -- we need to fix the "stobj" coming
out of the stobj-table if we want to prove something about the types
of that stobj's fields.