Fixed bug in guard generation for stobj-table field accesses in stobj-let bindings.  Slightly extended stobj-table tests.
Thanks to Sol Swords for an example that led me to discover this bug.

NOTE: Stobj-tables are still experimental.  Guard generation in the
presence of stobj-table accesses is now correct (I think), but Rob
Sumners and Sol Swords have convinced me to try to change it again to
be more useful.  We've agreed on a plan and I'm working on it.