Changed guard generation for stobj-table field accesses to avoid inserting stobj fixers.  Updated :doc on stobj-tables.  Control a bit of THM output with PROVE instead of SUMMARY output type.
Thanks to Sol Swords and Rob Sumners for feedback on issues with how
ACL2 was generating guard proof obligations that involve stobj-table
field accesses.  Formerly (though this was a recent change), stobj
fixers were applied to each such access before generating guards.
Now, stobj recognizer assumptions are added for such accesses, as
agreed with Sol and Rob.  (Implementation note: The key change was to
function guard-clauses, which now returns an "environment", as
discussed in a comment in that function.)

Quoting :doc note-8-5:

  The string "Proof succeeded" or "Proof skipped" was formerly printed
  when SUMMARY output is not inhibited (see
  [set-inhibit-output-lst]), but now that printing takes place
  instead when PROVE output is not inhibited.  Thanks to Pete
  Manolios for pointing out a printing issue that is resolved with
  this change.

Note that stobj fixers are still supported but are no longer mentioned
in :doc topics.  If they turn out not to be needed any longer, as I
rather expect, then they will probably be removed.